FF refactor for internal teams
